+++ src/core/prelude/string/const.ts
@@ -6,6 +6,8 @@
  * https://github.com/V4Fire/Core/blob/master/LICENSE
  */
 
+/* eslint-disable no-misleading-character-class */
+
 export const
 	capitalizeCache = Object.createDict<string>(),
 	camelizeCache = Object.createDict<string>(),
@@ -16,3 +18,12 @@ export const
 	isDigital = /\d/,
 	normalizeRgxp = /(^[\s_-]+)|([\s_-]+$)|([\s_-]+)/g,
 	camelizeRgxp = /(^[\s_-]+)|([\s_-]+$)|[\s_-]+([^\s-]|$)/g;
+
+export const unicode = {
+	zeroWidthJoiner: /\u200D/,
+	regionalIndicators: /\uD83C[\uDDE6-\uDDFF]/,
+	emojiWithColorModifiers: /[\u261D-\u26F9\u270A-\u270D]|\uD83C[\uDF85-\uDFFF]|\uD83D[\uDC00-\uDD96\uDE45-\uDE4F\uDEA3-\uDECC]|\uD83E[\uDD0F-\uDDDD]/,
+	textModifiers: /[\u0300-\u036F\u1AB0-\u1AFF\u20D0-\u20FF]/,
+	colorModifiers: /\uD83C[\uDFFB-\uDFFF]/,
+	modifiers: /[\u200D\uFE0F]/
+};