+++ src/core/async/modules/proxy/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.30.3 (2021-01-28)
+
+#### :house: Internal
+
+* Now all promises are wrapped by using `SyncPromise`
+
 ## v3.20.0 (2020-07-05)
 
 #### :rocket: New Feature