+++ src/core/prelude/string/CHANGELOG.md
@@ -9,6 +9,22 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.30.1 (2021-01-11)
+
+#### :bug: Bug Fix
+
+* Fixed iterations over chained emoji-s
+
+## v3.30.0 (2021-01-10)
+
+#### :rocket: New Feature
+
+* Added `letters` and `String.letters`
+
+#### :bug: Bug Fix
+
+* Improved Unicode support
+
 ## v3.20.0 (2020-07-05)
 
 #### :rocket: New Feature